+++ public/locale/en.json
@@ -1452,6 +1452,7 @@
   "discharge_confirmation_message": "You are about to mark this patient as discharged.",
   "discharge_confirmation_status_change": "The encounter status will be changed to 'Discharged'",
   "discharge_confirmation_summary_required": "You will have an option to provide discharge summary advice details",
+  "discharge_date_and_time": "Discharge Date and Time",
   "discharge_date_time": "Discharge Date and Time",
   "discharge_details": "Discharge Details",
   "discharge_disposition": "Discharge Disposition",
@@ -2450,6 +2451,7 @@
   "location_status__unknown": "Unknown",
   "location_updated": "Location Updated",
   "location_updated_successfully": "Location updated successfully",
+  "location_will_appear_here": "Patient location will appear here when assigned",
   "locations": "Locations",
   "locations_selected": "locations selected",
   "locations_under_my_care_for_immediate_transfer": "Locations under my care for immediate transfer",
@@ -2699,6 +2701,7 @@
   "next_sessions": "Next Sessions",
   "next_week_short": "Next wk",
   "no": "No",
+  "no_account_found": "No account found",
   "no_accounts_found": "No accounts found",
   "no_active_account_found": "No active account found",
   "no_active_beds_assigned": "No active beds assigned",
@@ -4172,7 +4175,7 @@
   "session_expired_message": "It looks like your session timed out for a moment. Take a quick breather, then log in again to continue.",
   "session_expired_msg": "It appears that your session has expired. This could be due to inactivity. Please login again to continue.",
   "session_slots_info": "{{slots}} slots of {{minutes}} mins.",
-  "session_slots_info_striked": " <s>{{intended_slots}} slots</s>  {{actual_slots}} slots of {{minutes}} mins.",
+  "session_slots_info_striked": "  {{intended_slots}} slots   {{actual_slots}} slots of {{minutes}} mins.",
   "session_title": "Session Title",
   "session_title_placeholder": "IP Rounds",
   "session_type": "Session Type",
@@ -4206,6 +4209,7 @@
   "show_all_slots": "Show all slots",
   "show_available_beds_only": "Show Available Beds Only",
   "show_default_presets": "Show Default Presets",
+  "show_less": "Show less",
   "show_notes": "Show notes",
   "show_on_map": "Show on Map",
   "show_patient_details_and_specimen_id": "Show patient details and specimen ID?",
