+++ src/pages/Encounters/tabs/overview/summary-panel-details-tab.tsx
@@ -1,0 +1,424 @@
+import { useQueryClient } from "@tanstack/react-query";
+import { format } from "date-fns";
+import { Plus, Signal, SquarePen } from "lucide-react";
+import { Link } from "raviger";
+import React from "react";
+import { useTranslation } from "react-i18next";
+
+import CareIcon from "@/CAREUI/icons/CareIcon";
+
+import { Badge } from "@/components/ui/badge";
+import { Button } from "@/components/ui/button";
+import { Separator } from "@/components/ui/separator";
+
+import { Avatar } from "@/components/Common/Avatar";
+import { QuestionnaireSearch } from "@/components/Questionnaire/QuestionnaireSearch";
+import TagAssignmentSheet from "@/components/Tags/TagAssignmentSheet";
+
+import { StatusBadge } from "@/pages/Encounters/EncounterProperties";
+import { useEncounter } from "@/pages/Encounters/utils/EncounterProvider";
+import {
+  ENCOUNTER_CLASSES_COLORS,
+  ENCOUNTER_CLASS_ICONS,
+  ENCOUNTER_PRIORITY_COLORS,
+  ENCOUNTER_STATUS_COLORS,
+  ENCOUNTER_STATUS_ICONS,
+} from "@/types/emr/encounter/encounter";
+import { getTagHierarchyDisplay } from "@/types/emr/tagConfig/tagConfig";
+
+import { Account } from "./summary-panel-details-tab/account";
+import { AuditLogs } from "./summary-panel-details-tab/auditlogs";
+import { DepartmentsAndTeams } from "./summary-panel-details-tab/department-and-team";
+import { DischargeDetails } from "./summary-panel-details-tab/discharge-summary";
+import { EmptyState } from "./summary-panel-details-tab/empty-state";
+import { HospitalizationDetails } from "./summary-panel-details-tab/hospitalisation";
+import { Locations } from "./summary-panel-details-tab/locations";
+import { ManageCareTeam } from "./summary-panel-details-tab/manage-care-team";
+import { Questionnaires } from "./summary-panel-details-tab/questionnaires";
+
+export const SummaryPanelDetailTab = () => {
+  const { t } = useTranslation();
+  const { selectedEncounter: encounter, canEditSelectedEncounter: canEdit } =
+    useEncounter();
+  const queryClient = useQueryClient();
+
+  if (!encounter) return null;
+  const EncounterClassIcon = ENCOUNTER_CLASS_ICONS[encounter.encounter_class];
+
+  return (
+    <div>
+      <div className="flex flex-col p-3 bg-white -mt-1 rounded-lg gap-4">
+        <div className="flex flex-row gap-14">
+          <div className="flex flex-col gap-4">
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("status")}:
+              </span>
+              <div>
+                <Badge variant={ENCOUNTER_STATUS_COLORS[encounter.status]}>
+                  {React.createElement(
+                    ENCOUNTER_STATUS_ICONS[encounter.status],
+                    {
+                      className: "size-4",
+                    },
+                  )}
+                  {t(`encounter_status__${encounter.status}`)}
+                </Badge>
+              </div>
+            </div>
+
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("encounter_class")}:
+              </span>
+              <div>
+                <Badge
+                  variant={ENCOUNTER_CLASSES_COLORS[encounter.encounter_class]}
+                >
+                  <EncounterClassIcon className="size-3" />
+                  <span className="whitespace-nowrap">
+                    {t(`encounter_class__${encounter.encounter_class}`)}
+                  </span>
+                </Badge>
+              </div>
+            </div>
+
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("location")}:
+              </span>
+              <div>
+                <Badge
+                  variant="secondary"
+                  className="inline-flex items-center gap-1.5"
+                >
+                  <CareIcon icon="l-location-point" className="size-3" />
+                  {encounter.current_location?.name || t("none")}
+                </Badge>
+              </div>
+            </div>
+          </div>
+
+          <div className="flex flex-col gap-4">
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("priority")}:
+              </span>
+              <div>
+                <Badge variant={ENCOUNTER_PRIORITY_COLORS[encounter.priority]}>
+                  <Signal className="size-3" />
+                  {t(`encounter_priority__${encounter.priority}`)}
+                </Badge>
+              </div>
+            </div>
+
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("dep_and_teams")}:
+              </span>
+              <div className="flex flex-wrap gap-2">
+                {encounter.organizations.length > 0 ? (
+                  <>
+                    {encounter.organizations.map((org) => (
+                      <Badge key={org.id} variant="blue" className="capitalize">
+                        {org.name}
+                      </Badge>
+                    ))}
+                  </>
+                ) : (
+                  <p className="text-sm text-gray-500">
+                    {t("no_departments_assigned")}
+                  </p>
+                )}
+              </div>
+            </div>
+
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("hospitalisation")}:
+              </span>
+              <div>
+                <Badge variant="blue">Re-admission</Badge>
+              </div>
+            </div>
+          </div>
+        </div>
+        <div>
+          <span className="text-sm font-medium text-gray-700">
+            {t("encounter_tags")}:
+          </span>
+          <div className="flex flex-wrap gap-2">
+            {encounter.tags.length > 0 ? (
+              <>
+                {encounter.tags.map((tag) => (
+                  <Badge
+                    key={tag.id}
+                    variant="secondary"
+                    className="capitalize"
+                    title={tag.description}
+                  >
+                    {tag.display}
+                  </Badge>
+                ))}
+              </>
+            ) : (
+              <p className="text-sm text-gray-500">{t("no_tags")}</p>
+            )}
+          </div>
+        </div>
+        <Separator />
+        <div className="flex flex-row -mt-2 gap-5">
+          <div className="flex flex-col gap-2">
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("start_date")}:
+              </span>
+              <div className="text-sm text-gray-950 font-semibold">
+                {encounter.period.start
+                  ? format(encounter.period.start, "dd MMM yyyy hh:mma")
+                  : ""}
+              </div>
+            </div>
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("hospital_identifier")}
+              </span>
+              <div className="text-sm text-gray-950 font-semibold">
+                {encounter.external_identifier || "--"}
+              </div>
+            </div>
+
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("account")}:
+              </span>
+              <div className="text-sm text-gray-950 font-semibold">
+                {encounter.external_identifier || "BILL-ENC-78934"}
+              </div>
+            </div>
+          </div>
+
+          <div className="flex flex-col gap-2">
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("end_date")}:
+              </span>
+              <div className="text-sm text-gray-950 font-semibold">
+                {encounter.period.end
+                  ? format(encounter.period.end, "dd MMM yyyy hh:mma")
+                  : "(Ongoing)"}
+              </div>
+            </div>
+
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("patient_id_abha")}:
+              </span>
+              <div className="text-sm text-gray-950 font-semibold">--</div>
+            </div>
+
+            <div>
+              <span className="text-sm font-medium text-gray-700">
+                {t("care_team")}:
+              </span>
+              <div className="flex items-center gap-2">
+                <div className="flex -space-x-2">
+                  {encounter.care_team.slice(0, 3).map((member) => (
+                    <Avatar
+                      key={member.member.id}
+                      name={member.member.first_name}
+                      imageUrl={member.member.profile_picture_url}
+                      className="size-8 rounded-full"
+                    />
+                  ))}
+                </div>
+                {encounter.care_team.length > 3 && (
+                  <span className="text-sm font-medium text-gray-600">
+                    +{encounter.care_team.length - 3}
+                  </span>
+                )}
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+      <div className="@xs:hidden xl:flex flex-col gap-4">
+        <div className="flex flex-wrap gap-2 border bg-gray-100 border-gray-200 rounded-md p-1">
+          <div className="flex items-center justify-between w-full p-2 text-gray-950">
+            <span className="font-semibold ">{t("encounter_details")}</span>
+            <Button variant="ghost" size="xs" asChild>
+              <Link
+                href={`/facility/${encounter.facility.id}/patient/${encounter.patient.id}/encounter/${encounter.id}/questionnaire/encounter`}
+              >
+                <SquarePen
+                  className="size-4 text-gray-950 cursor-pointer"
+                  strokeWidth={1.5}
+                />
+              </Link>
+            </Button>
+          </div>
+          <div className="flex flex-wrap gap-2 justify-between bg-white w-full p-2 rounded-md shadow">
+            <div className="flex flex-col gap-1">
+              <span className="text-sm font-medium">{t("status")}: </span>
+              <div>
+                <StatusBadge encounter={encounter} />
+              </div>
+            </div>
+            <div className="flex flex-col gap-1">
+              <span className="text-sm font-medium">
+                {t("encounter_class")}:{" "}
+              </span>
+              <div>
+                <Badge variant="teal" size="sm">
+                  <EncounterClassIcon className="size-3" />
+                  <span className="whitespace-nowrap">
+                    {t(`encounter_class__${encounter.encounter_class}`)}
+                  </span>
+                </Badge>
+              </div>
+            </div>
+            <div className="flex flex-col gap-1">
+              <span className="text-sm font-medium">{t("priority")}: </span>
+              <div>
+                <Badge variant="orange" size="sm">
+                  <span className="whitespace-nowrap">
+                    {t(`encounter_priority__${encounter.priority}`)}
+                  </span>
+                </Badge>
+              </div>
+            </div>
+            <Separator className="my-2" />
+            <div className="hidden md:flex flex-col gap-1">
+              <div>
+                <span className="text-sm font-medium text-gray-700">
+                  {t("start_date")}:
+                </span>
+                <div className="text-sm text-gray-950 font-semibold">
+                  {encounter.period.start ? (
+                    <>
+                      {format(encounter.period.start, "dd MMM yyyy")}
+                      <div className="text-gray-600">
+                        {format(encounter.period.start, "hh:mma")}
+                      </div>
+                    </>
+                  ) : (
+                    <span>--</span>
+                  )}
+                </div>
+              </div>
+            </div>
+
+            <div className="hidden md:flex flex-col gap-1">
+              <div>
+                <span className="text-sm font-medium text-gray-700">
+                  {t("end_date")}:
+                </span>
+                <div className="text-sm text-gray-950 font-semibold">
+                  {encounter.period.end ? (
+                    <>
+                      {format(encounter.period.end, "dd MMM yyyy")},
+                      <div className="text-gray-600">
+                        {format(encounter.period.end, "hh:mma")}
+                      </div>
+                    </>
+                  ) : (
+                    <span>--({t("ongoing")})</span>
+                  )}
+                </div>
+              </div>
+            </div>
+          </div>
+          <Button variant="outline" className="w-full" asChild>
+            <Link
+              href={`/facility/${encounter.facility.id}/patient/${encounter.patient.id}/encounter/${encounter.id}/questionnaire/encounter`}
+            >
+              <SquarePen className="size-3 text-gray-950" strokeWidth={1.5} />
+              <span className="text-gray-950">{t("update_details")}</span>
+            </Link>
+          </Button>
+        </div>
+        <div className="bg-gray-100 rounded-md border border-gray-200 p-1">
+          <div className="flex items-center justify-between w-full">
+            <span className="font-semibold text-gray-950 p-2">
+              {t("encounter_tags")}
+            </span>
+            {canEdit && (
+              <TagAssignmentSheet
+                entityType="encounter"
+                entityId={encounter.id}
+                currentTags={encounter.tags}
+                onUpdate={() => {
+                  queryClient.invalidateQueries({
+                    queryKey: ["encounter", encounter.id],
+                  });
+                }}
+                trigger={
+                  <Button variant="ghost" size="xs">
+                    <SquarePen
+                      className="size-3 text-gray-950"
+                      strokeWidth={1.5}
+                    />
+                  </Button>
+                }
+                canWrite={canEdit}
+              />
+            )}
+          </div>
+          <div className="flex flex-wrap bg-white w-full p-2 rounded-md gap-2 shadow">
+            {encounter.tags.length > 0 ? (
+              <>
+                {encounter.tags.map((tag) => (
+                  <Badge
+                    key={tag.id}
+                    variant="secondary"
+                    className="capitalize"
+                    title={tag.description}
+                  >
+                    {getTagHierarchyDisplay(tag)}
+                  </Badge>
+                ))}
+              </>
+            ) : (
+              <EmptyState message={t("no_tags")} />
+            )}
+          </div>
+        </div>
+        <div className="bg-gray-100 rounded-md p-2 border border-gray-200">
+          <div className="flex items-center justify-between w-full p-2">
+            <span className="font-semibold text-gray-950 ">{t("forms")}</span>
+            <QuestionnaireSearch
+              trigger={
+                <Button variant="ghost" size="xs">
+                  <Plus className="size-3 text-gray-950" />
+                </Button>
+              }
+              subjectType="encounter"
+            />
+          </div>
+          {canEdit && <Questionnaires encounter={encounter} />}
+        </div>
+        <div className="bg-gray-100 rounded-md w-full border border-gray-200">
+          <Locations canEdit={canEdit} encounter={encounter} />
+        </div>
+        <div className="bg-gray-100 rounded-md w-full border border-gray-200">
+          <ManageCareTeam />
+        </div>
+        <div className="bg-gray-100 rounded-md w-full border border-gray-200">
+          <DepartmentsAndTeams canEdit={canEdit} encounter={encounter} />
+        </div>
+        <div className="bg-gray-100 rounded-md w-full border border-gray-200">
+          <DischargeDetails encounter={encounter} />
+        </div>
+        <div className="bg-gray-100 rounded-md w-full border border-gray-200">
+          <HospitalizationDetails encounter={encounter} />
+        </div>
+        <div className="bg-gray-100 rounded-md w-full border border-gray-200">
+          <Account encounter={encounter} canEdit={canEdit} />
+        </div>
+        <div>
+          <AuditLogs encounter={encounter} />
+        </div>
+      </div>
+    </div>
+  );
+};
