+++ src/pages/Encounters/tabs/overview/summary-panel-details-tab/manage-care-team.tsx
@@ -1,0 +1,85 @@
+import { NotebookPen, SquarePen } from "lucide-react";
+import { useTranslation } from "react-i18next";
+
+import { Badge } from "@/components/ui/badge";
+import { Button } from "@/components/ui/button";
+
+import { CareTeamSheet } from "@/components/CareTeam/CareTeamSheet";
+import { Avatar } from "@/components/Common/Avatar";
+
+import { useEncounter } from "@/pages/Encounters/utils/EncounterProvider";
+
+export const ManageCareTeam = () => {
+  const { t } = useTranslation();
+  const {
+    selectedEncounter: encounter,
+    selectedEncounterPermissions: { canWriteEncounter: canWrite },
+  } = useEncounter();
+
+  if (!encounter) {
+    return (
+      <Button disabled variant="outline" className="justify-start">
+        <NotebookPen />
+        {t("manage_care_team")}
+      </Button>
+    );
+  }
+
+  return (
+    <CareTeamSheet
+      encounter={encounter}
+      trigger={
+        <div className="bg-gray-100 rounded-md">
+          <div className="flex justify-between items-center p-2 w-full">
+            <span className="text-black font-semibold">
+              {canWrite ? t("manage_care_team") : t("view_care_team")}
+            </span>
+            <SquarePen className="size-4 cursor-pointer" />
+          </div>
+          <div className="bg-white p-2 rounded-md mx-1">
+            {encounter.care_team.length > 0 ? (
+              <div className="flex flex-col gap-1">
+                {encounter.care_team.slice(0, 3).map((member, index) => (
+                  <div
+                    key={member.member.id}
+                    className="flex items-center gap-2 p-2 rounded-md border border-gray-100 bg-gray-200/20"
+                  >
+                    <Avatar
+                      key={member.member.id}
+                      name={member.member.first_name}
+                      imageUrl={member.member.profile_picture_url}
+                      className="size-9 rounded-full"
+                    />{" "}
+                    <div className="flex items-center justify-between w-full">
+                      <div className="flex flex-col">
+                        <span className="font-medium text-black text-sm">
+                          {member.member.first_name}
+                        </span>
+                        <span className="text-xs text-gray-500">
+                          {member.member.user_type}
+                        </span>
+                      </div>
+                      {index === 0 && (
+                        <Badge variant="primary" className="font-normal">
+                          {t("primary")}
+                        </Badge>
+                      )}
+                    </div>
+                  </div>
+                ))}
+                {encounter.care_team.length > 3 && (
+                  <span className="text-sm font-medium text-black underline">
+                    +{encounter.care_team.length - 3} {t("members")}
+                  </span>
+                )}
+              </div>
+            ) : (
+              <p className="text-sm text-gray-500">{t("no_care_team")}</p>
+            )}
+          </div>
+        </div>
+      }
+      canWrite={canWrite}
+    />
+  );
+};
